+++ src/components/InputField/InputField.stories.tsx
@@ -17,7 +17,7 @@ const meta: Meta<typeof InputField> = {
   },
   args: {
     className: 'w-[384px]',
-    sublabel: 'Additional descriptive text for the field.',
+    subLabel: 'Additional descriptive text for the field.',
   },
   argTypes: {
     type: {
@@ -47,12 +47,12 @@ type Story = StoryObj<typeof InputField>;
 export const Default: Story = {
   args: {
     label: 'Default input field',
-    sublabel: '',
+    subLabel: '',
   },
 };
 
 /**
- * An input field can have both a footnote (describing how the field is validated), and a sublabel
+ * An input field can have both a footnote (describing how the field is validated), and a subLabel
  * (a longer description to elaborate on the label)
  */
 export const WithFullDescription: Story = {